Improve key parsing when obs are not the same
mu.pl.scatter uses the new key parsing and expansion (d5a15a0).

Now it is also possible to use the same parsing backend
and consequently the same plotting functions
for MuData objects where some modalities have observations
that are not present in other modalities.

This implementation relies on .obsmap to achieve that.